ci: run ruff once so a miss is one email, not eight
The eight-way matrix used to fail eight jobs on lint and never run pytest. Hung Qt tests sat until GitHub's six-hour wall. Pre-commit now catches ruff before the push.

+++ arelis/physics/elements.py
@@ -116,7 +116,7 @@ def well_mesh_indices(n: int) -> list[int]:
 def well_inner_ring(n: int) -> int:
     """Skip rings that sit on the photosphere. Default span is 4 R."""
     n_r = max(int(n), 4)
-    return max(2, int(round(n_r * 0.28)))
+    return max(2, round(n_r * 0.28))
 
 
 def well_line_indices(n: int) -> list[int]:
